dynamic near far
