+++ index.html
@@ -177,7 +177,7 @@
       addProjectElement(
         'Local Avatar Game',
         divPrjUDGame,
-        './examples/Avatar.html',
+        './examples/AvatarGame.html',
         'This example implements a local mini game illustrating an avatar walking in the city',
         './examples/assets/img/examples_images/AvatarGame.png'
       );
